# Docs/API/openapi.yaml
openapi: 3.0.3
info:
  title: HRM System API
  description: >
    REST API for the full HRM system as currently analyzed: Employee Profile,
    Organization Management, Salary Master Data, Salary Grade Promotion, and
    Contract Management. Designed from the current User Stories / Use Cases,
    Information Architecture / Screens Hierarchy / UI-UX, the C4 model
    (`Docs/c4/`), and the database design (`Docs/Database/HRM_System.dbml`).

    Implementation status: all 57 endpoints across all 11 tags, including
    **Contracts**, are implemented in `backend/`; see `backend/README.md`.
    The backend has no authentication yet (`bearerAuth` is declared for
    documentation only), serves routes without a version prefix, and the
    server URL below is a placeholder. See `Docs/API/README.md`.
  version: "0.3.0"
servers:
  - url: https://api.example.com/v1
...
security:
  - bearerAuth: []
paths:
...
  /contracts:
    post:
      tags: [Contracts]
      operationId: createContract
      summary: Create a labor contract for an employee
      description: >
        US-CON-01 — employeeId, contractType, contractNumber, startDate, and
        contractSalaryAmount are required; contractNumber must be unique
        (BR-CON-01, 02). status is the initial status, Draft or Active
        (BR-CON-08).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractRequest"
      responses:
        "201":
          description: Contract created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDetail"
        "400":
          description: >
            endDate does not match contractType (BR-CON-04), endDate is not
            later than startDate (BR-CON-05), or contractSalaryAmount is zero
            or negative (BR-CON-06)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: >
            contractNumber already in use (BR-CON-01), the employee's
            employmentStatus is Terminated (BR-CON-07), or status is Active
            while startDate is later than today or the employee already has
            another Active contract (BR-CON-09)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [Contracts]
      operationId: listContracts
      summary: Search, filter, and list contracts, including contracts expiring soon
      description: >
        US-CON-02 — search matches employeeCode, employee fullName, or
        contractNumber; filters may be combined; contracts of every status
        are included unless status is given (BR-CON-11–14). US-CON-05 —
        set expiringSoon to true to instead list the Active contracts ending
        within window days (30 or 60, default 30) plus any overdue Active
        contract, ordered by endDate ascending; this can be combined with
        search and the other filters (BR-CON-25–27, 32). Because that view
        only ever contains Active contracts, combining expiringSoon with a
        status other than ACTIVE returns no results.
      parameters:
        - name: search
          in: query
          description: Matches employee code, employee full name, or contract number.
          schema:
            type: string
        - name: contractType
          in: query
          schema:
            $ref: "#/components/schemas/ContractType"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ContractStatus"
        - $ref: "#/components/parameters/FromDate"
        - $ref: "#/components/parameters/ToDate"
        - name: expiringSoon
          in: query
          description: US-CON-05 — when true, list only expiring-soon and overdue Active contracts.
          schema:
            type: boolean
            default: false
        - name: window
          in: query
          description: Expiring-soon window in days, ignored unless expiringSoon is true (BR-CON-25).
          schema:
            type: integer
            enum: [30, 60]
            default: 30
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A page of contracts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractPage"
  /contracts/{contractId}:
    get:
      tags: [Contracts]
      operationId: getContract
      summary: Get contract detail
      description: >
        US-CON-03 — includes the employee's currently recorded code,
        fullName, Organizational Unit, Job Title, and employmentStatus, not a
        copy frozen at contract creation (BR-CON-15). terminationDate and
        terminationReason are set only when status is Terminated.
      parameters:
        - $ref: "#/components/parameters/ContractId"
      responses:
        "200":
          description: Contract detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Contracts]
      operationId: updateContract
      summary: Update a Draft contract
      description: >
        US-CON-06 — only allowed while status is Draft (BR-CON-28); employeeId
        and status are not editable here — delete this contract and create a
        new one if the wrong employee was selected (BR-CON-29). The same
        validation rules as create apply to the updated fields (BR-CON-30).
      parameters:
        - $ref: "#/components/parameters/ContractId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContractRequest"
      responses:
        "200":
          description: Contract updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDetail"
        "400":
          description: >
            endDate does not match contractType, endDate is not later than
            startDate, or contractSalaryAmount is zero or negative
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Contract is not Draft, or the changed contractNumber is already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Contracts]
      operationId: deleteContract
      summary: Delete a Draft contract
      description: >
        US-CON-06 — only allowed while status is Draft (BR-CON-28). The
        contract is permanently removed and its contractNumber can be reused
        (BR-CON-31) — the only hard delete in this API.
      parameters:
        - $ref: "#/components/parameters/ContractId"
      responses:
        "204":
          description: Contract deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Contract is not Draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /contracts/{contractId}/activate:
    post:
      tags: [Contracts]
      operationId: activateContract
      summary: Activate a Draft contract
      description: >
        US-CON-04 — only allowed while status is Draft, startDate is today or
        earlier, the employee has no other Active contract, and the
        employee's employmentStatus is not Terminated (BR-CON-18).
      parameters:
        - $ref: "#/components/parameters/ContractId"
      responses:
        "200":
          description: Contract activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            Contract is not Draft, startDate is later than today, the
            employee already has another Active contract, or the employee's
            employmentStatus is Terminated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /contracts/{contractId}/expire:
    post:
      tags: [Contracts]
      operationId: expireContract
      summary: Mark an Active contract as Expired
      description: >
        US-CON-04 — only allowed while status is Active, endDate is set, and
        endDate is earlier than today (BR-CON-19).
      parameters:
        - $ref: "#/components/parameters/ContractId"
      responses:
        "200":
          description: Contract marked as Expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDetail"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Contract is not Active, has no endDate, or endDate is today or later
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /contracts/{contractId}/terminate:
    post:
      tags: [Contracts]
      operationId: terminateContract
      summary: Terminate an Active contract before its normal expiry
      description: >
        US-CON-04 — terminationDate and terminationReason are required;
        terminationDate must not be earlier than startDate and, if endDate is
        set, not later than it (BR-CON-21, 22).
      parameters:
        - $ref: "#/components/parameters/ContractId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerminateContractRequest"
      responses:
        "200":
          description: Contract terminated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDetail"
        "400":
          description: terminationDate or terminationReason is missing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Contract is not Active, or terminationDate is outside the contract term
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT issued after login (see Arc42 ADR-07). The token-issuing login
        endpoint itself is not yet designed in this spec.
  parameters:
...
    ContractId:
      name: contractId
      in: path
      required: true
      schema:
        type: integer
        format: int32
    FromDate:
      name: fromDate
      in: query
      schema:
        type: string
        format: date
    ToDate:
      name: toDate
      in: query
      schema:
        type: string
        format: date
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: reason is required when rejecting an employee
...
    EmploymentStatus:
      type: string
      enum: [ACTIVE, ON_LEAVE, TERMINATED]
...
    ContractType:
      type: string
      enum: [PROBATION, FIXED_TERM, INDEFINITE_TERM]
    ContractStatus:
      type: string
      description: >
        DRAFT: recorded but not yet in effect. ACTIVE: in effect. EXPIRED /
        TERMINATED: final, read-only (BR-CON-16). Transitions are explicit
        HR Staff actions, never date-triggered (BR-CON-17).
      enum: [DRAFT, ACTIVE, EXPIRED, TERMINATED]
    CreateContractRequest:
      type: object
      required: [employeeId, contractType, contractNumber, startDate, contractSalaryAmount, status]
      properties:
        employeeId:
          type: integer
          format: int32
        contractType:
          $ref: "#/components/schemas/ContractType"
        contractNumber:
          type: string
          example: CT-2026-010
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
          description: Required for PROBATION/FIXED_TERM, must be omitted for INDEFINITE_TERM (BR-CON-04).
        contractSalaryAmount:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
        salaryNote:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ContractStatus"
          description: Initial status — DRAFT or ACTIVE only (BR-CON-08).
    UpdateContractRequest:
      type: object
      required: [contractType, contractNumber, startDate, contractSalaryAmount]
      description: Only allowed while the contract is DRAFT. employeeId and status are not editable here (BR-CON-29, 30).
      properties:
        contractType:
          $ref: "#/components/schemas/ContractType"
        contractNumber:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        contractSalaryAmount:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
        salaryNote:
          type: string
          nullable: true
    TerminateContractRequest:
      type: object
      required: [terminationDate, terminationReason]
      properties:
        terminationDate:
          type: string
          format: date
        terminationReason:
          type: string
    Contract:
      type: object
      properties:
        id:
          type: integer
          format: int32
        employeeId:
          type: integer
          format: int32
        employeeCode:
          type: string
          readOnly: true
        employeeFullName:
          type: string
          readOnly: true
        contractNumber:
          type: string
        contractType:
          $ref: "#/components/schemas/ContractType"
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        contractSalaryAmount:
          type: number
          format: double
        salaryNote:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ContractStatus"
        overdue:
          type: boolean
          readOnly: true
          description: True when status is ACTIVE and endDate is earlier than today (BR-CON-20).
        terminationDate:
          type: string
          format: date
          nullable: true
        terminationReason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ContractDetail:
      allOf:
        - $ref: "#/components/schemas/Contract"
        - type: object
          properties:
            organizationalUnitName:
              type: string
              readOnly: true
            jobTitleName:
              type: string
              readOnly: true
            employmentStatus:
              $ref: "#/components/schemas/EmploymentStatus"
    ContractPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Contract"
        page:
          type: integer
        pageSize:
          type: integer
        totalItems:
          type: integer